refactor: user_dict 機能をサブサーバに切り出して整理
